backend_django/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["mat <mateusz.baranczyk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
markdown = "^3.6"
djangorestframework-simplejwt = "^5.3.1"
mysql-connector-python = "^9.0.0"
requests = "^2.32.3"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.0.8"
django-redis-session-store = "^0.1.1"
gunicorn = "^23.0.0"

[tool.poetry.dev-dependencies]
black = "^24.4.2"
flake8 = "^7.1.0"
flake8-pyproject = "^1.2.3"
isort = "^5.13.2"
pytest = "^7.2.0"

[tool.isort]
profile = "black"
line_length = 79

[tool.black]
line-length = 79
target_version = ['py311']

[tool.flake8]
ignore = ['E501']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
